// src/types/models/AccountType.ts
/**
 * Yapily API
 * The Yapily API enables connections between your application and users\' banks. For more information check out our [documentation](https://docs.yapily.com/).<br><br>In particular, make sure to view our [Getting Started](https://docs.yapily.com/pages/home/getting-started/) steps if this is your first time here.<br><br>While testing the API, our list of [sandbox credentials](https://docs.yapily.com/pages/key-concepts/sandbox-credentials/) maybe useful.
 *
 * OpenAPI spec version: 2.30.1
 * Contact: support@yapily.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

/**
* The type of account e.g. (Credit Card, Savings).
*/
export enum AccountType {
    CashTrading = 'CASH_TRADING',
    CashIncome = 'CASH_INCOME',
    CashPayment = 'CASH_PAYMENT',
    ChargeCard = 'CHARGE_CARD',
    Charges = 'CHARGES',
    Commission = 'COMMISSION',
    CreditCard = 'CREDIT_CARD',
    Current = 'CURRENT',
    EMoney = 'E_MONEY',
    LimitedLiquiditySavingsAccount = 'LIMITED_LIQUIDITY_SAVINGS_ACCOUNT',
    Loan = 'LOAN',
    MarginalLending = 'MARGINAL_LENDING',
    MoneyMarket = 'MONEY_MARKET',
    Mortgage = 'MORTGAGE',
    NonResidentExternal = 'NON_RESIDENT_EXTERNAL',
    Other = 'OTHER',
    Overdraft = 'OVERDRAFT',
    OvernightDeposit = 'OVERNIGHT_DEPOSIT',
    PrepaidCard = 'PREPAID_CARD',
    Salary = 'SALARY',
    Savings = 'SAVINGS',
    Settlement = 'SETTLEMENT',
    Tax = 'TAX',
    Unknown = 'UNKNOWN'
}