// src/types/models/ConsolidatedAccountInformation.ts
/**
 * Yapily API
 * The Yapily API enables connections between your application and users\' banks. For more information check out our [documentation](https://docs.yapily.com/).<br><br>In particular, make sure to view our [Getting Started](https://docs.yapily.com/pages/home/getting-started/) steps if this is your first time here.<br><br>While testing the API, our list of [sandbox credentials](https://docs.yapily.com/pages/key-concepts/sandbox-credentials/) maybe useful.
 *
 * OpenAPI spec version: 2.30.1
 * Contact: support@yapily.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { AccountBalance } from '../models/AccountBalance';
import { HttpFile } from '../http/http';

/**
* Summary information regarding account balances of the overall account provided by the bank
*/
export class ConsolidatedAccountInformation {
    /**
    * Identifier of the consolidated account. When used in Get Account Transactions calls, the transactions between the sub-accounts will not be reported
    */
    'id'?: string;
    'accountBalances'?: Array<AccountBalance>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "string",
            "format": ""
        },
        {
            "name": "accountBalances",
            "baseName": "accountBalances",
            "type": "Array<AccountBalance>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return ConsolidatedAccountInformation.attributeTypeMap;
    }

    public constructor() {
    }
}